refactor(agent-task): delegate controller execution policy to reusable core service (#4303)
