+++ src/core/agent_tasks.rs
@@ -139,7 +139,8 @@ pub use super::agent_task_provider::{
 /// Durable controller execution service entry points and report contracts.
 pub mod controller_service {
     pub use super::super::agent_task_controller_service::{
-        apply_event, init, init_from_spec, list, mark_human_ready, optional_bool, optional_string,
+        apply_event, apply_spec_dispatch_defaults, apply_spec_dispatch_defaults_with_cwd, init,
+        init_from_spec, list, mark_human_ready, optional_bool, optional_string,
         optional_string_array, optional_u32, optional_usize, plan_from_controller_request,
         plan_from_spec, resume, run_action, run_next, status, AgentTaskRepoLoopSpec,
         AgentTaskRepoLoopSpecAbility, AgentTaskRepoLoopSpecAgent, AgentTaskRepoLoopSpecArtifact,
